BaseInterceptor renamed to InterceptorBase
